fix(re_daemon): eliminate shutdown stall when no tasks are running
Three changes to prevent uvicorn from blocking on "Waiting for background
tasks to complete" during daemon shutdown:

1. McpStdioClient.close(): await cancelled reader tasks instead of
   leaving them in cancelling state, which uvicorn's background task
   tracker waits on indefinitely.

2. AgentManager.close(): terminate idle Pi child processes directly
   (SIGTERM) instead of sending a polite abort RPC and hoping the
   child exits voluntarily.  For processes with no in-progress tasks
   there is nothing to checkpoint, and idle processes may not drain
   stdin promptly.

3. GhidraAdapter.close(): wrap the close_database MCP call in a 5s
   timeout.  The previous 120s default timeout could stall shutdown
   for two minutes if the Ghidra MCP server was unresponsive.

+++ src/decompiled_cpp/shared/stubs_impl.cpp
@@ -49,7 +49,7 @@ unsigned int CRT_time(unsigned int* t) { return (unsigned int)time((time_t*)t); 
 
 /* ---- Game globals ---- */
 uint32_t g_game_time = 0;
-uint8_t  g_game_mode = 0;
+int32_t  g_game_mode = 0;  /* 0x4851F4: dword, read/written by CGWND_SetMode */
 char     g_empty_string = 0;
 char     g_install_path[256] = ".";
 void*    g_main_window = nullptr;
